fix(fsutil): bound AtomicWriteNew temp names
AtomicWriteNew still derived its temp name from the full basename, so
the ENAMETOOLONG fix only covered one of the two writers.

The umask tests use syscall.Umask, which does not build on Windows.
Moved behind the //go:build unix tag this package already uses for
lock_unix_test.go.

The gate's "test file" case wrote a plain .go fixture, so it never
exercised the claim that test files are scanned. It now writes a real
_test.go file — verified by excluding test files from the gate and
watching that case fail.
